refactor(proposal): PHPDoc removed and refactor for Proposal dirs.
