+++ app/Http/Requests/Quotes/Proposal/CreateQuoteProposalRequest.php
@@ -1,24 +1,16 @@
 <?php
 
-namespace App\Http\Requests\QuoteProposal;
+namespace App\Http\Requests\Quotes\Proposal;
 
 use Illuminate\Foundation\Http\FormRequest;
 
 class CreateQuoteProposalRequest extends FormRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
     public function authorize(): bool
     {
         return true;
     }
 
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array|string>
-     */
     public function rules(): array
     {
         return [
