+++ routes/api/quote_proposals.php
@@ -1,10 +1,9 @@
 <?php
 
 
-use App\Constants\UserTypesAbilitiesConstant;
 use App\Http\Controllers\QuoteProposal\CreateQuoteProposalController;
-use App\Http\Controllers\QuoteProposal\GetAllQuoteProposalsController;
-use App\Http\Controllers\QuoteProposal\ReviewQuoteProposalcontroller;
+use App\Http\Controllers\Quotes\Proposal\GetAllQuoteProposalsController;
+use App\Http\Controllers\Quotes\Proposal\ReviewQuoteProposalcontroller;
 
 
 Route::post('/', CreateQuoteProposalController::class)
